Moved SyncEvents test to TestCPU6502 unit-tests (VS2008 only)
