+++ test/TestCPU6502/TestCPU6502.cpp
@@ -3,10 +3,12 @@
 #include "../../source/Applewin.h"
 #include "../../source/CPU.h"
 #include "../../source/Memory.h"
+#include "../../source/SynchronousEventManager.h"
 
 // From Applewin.cpp
 bool g_bFullSpeed = false;
 enum AppMode_e g_nAppMode = MODE_RUNNING;
+SynchronousEventManager g_SynchronousEventMgr;
 
 // From Memory.cpp
 LPBYTE         memwrite[0x100];		// TODO: Init
@@ -27,8 +29,7 @@ iofunction		IOWrite[256] = {0};	// TODO: Init
 
 regsrec regs;
 
-static const int IRQ_CHECK_TIMEOUT = 128;
-static signed int g_nIrqCheckTimeout = IRQ_CHECK_TIMEOUT;
+bool g_irqOnLastOpcodeCycle = false;
 
 static eCpuType g_ActiveCPU = CPU_65C02;
 
@@ -1253,6 +1254,62 @@ int GH321_test()
 
 //-------------------------------------
 
+int testCB(int id, int underflowCycles, ULONG uExecutedCycles)
+{
+	return 0;
+}
+
+int SyncEvents_test(void)
+{
+	SyncEvent syncEvent0(0, 0x10, testCB);
+	SyncEvent syncEvent1(1, 0x20, testCB);
+	SyncEvent syncEvent2(2, 0x30, testCB);
+	SyncEvent syncEvent3(3, 0x40, testCB);
+
+	g_SynchronousEventMgr.Insert(&syncEvent0);
+	g_SynchronousEventMgr.Insert(&syncEvent1);
+	g_SynchronousEventMgr.Insert(&syncEvent2);
+	g_SynchronousEventMgr.Insert(&syncEvent3);
+	// id0 -> id1 -> id2 -> id3
+	if (syncEvent0.m_cyclesRemaining != 0x10) return 1;
+	if (syncEvent1.m_cyclesRemaining != 0x10) return 1;
+	if (syncEvent2.m_cyclesRemaining != 0x10) return 1;
+	if (syncEvent3.m_cyclesRemaining != 0x10) return 1;
+
+	g_SynchronousEventMgr.Remove(1);
+	g_SynchronousEventMgr.Remove(3);
+	g_SynchronousEventMgr.Remove(0);
+	if (syncEvent2.m_cyclesRemaining != 0x30) return 1;
+	g_SynchronousEventMgr.Remove(2);
+
+	//
+
+	syncEvent0.m_cyclesRemaining = 0x40;
+	syncEvent1.m_cyclesRemaining = 0x30;
+	syncEvent2.m_cyclesRemaining = 0x20;
+	syncEvent3.m_cyclesRemaining = 0x10;
+
+	g_SynchronousEventMgr.Insert(&syncEvent0);
+	g_SynchronousEventMgr.Insert(&syncEvent1);
+	g_SynchronousEventMgr.Insert(&syncEvent2);
+	g_SynchronousEventMgr.Insert(&syncEvent3);
+	// id3 -> id2 -> id1 -> id0
+	if (syncEvent0.m_cyclesRemaining != 0x10) return 1;
+	if (syncEvent1.m_cyclesRemaining != 0x10) return 1;
+	if (syncEvent2.m_cyclesRemaining != 0x10) return 1;
+	if (syncEvent3.m_cyclesRemaining != 0x10) return 1;
+
+	g_SynchronousEventMgr.Remove(3);
+	g_SynchronousEventMgr.Remove(0);
+	g_SynchronousEventMgr.Remove(1);
+	if (syncEvent2.m_cyclesRemaining != 0x20) return 1;
+	g_SynchronousEventMgr.Remove(2);
+
+	return 0;
+}
+
+//-------------------------------------
+
 int _tmain(int argc, _TCHAR* argv[])
 {
 	int res = 1;
@@ -1275,6 +1332,9 @@ int _tmain(int argc, _TCHAR* argv[])
 	if (res) return res;
 
 	res = GH292_test();
+	if (res) return res;
+
+	res = SyncEvents_test();
 	if (res) return res;
 
 	return 0;
